requirements.txt
-f /cvmfs/soft.computecanada.ca/custom/python/wheelhouse/avx2
-f /cvmfs/soft.computecanada.ca/custom/python/wheelhouse/generic
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.0.3
numpy==1.16.3
opencv-python==3.4.4.19
Pillow==5.3.0
pyparsing==2.4.0
python-dateutil==2.8.0
PyYAML==3.13
scikit-learn==0.20.2
scipy==1.2.1
six==1.12.0
torch==1.1.0
torch-gpu==1.0.0
torchvision==0.3.0
xmltodict==0.11.0
